feat(game/play): restructure play page - separate components
